Dashboard left side bar and right content part rendering
created functions addList and addListSave, showListByIdWithTasks. In class Model created fetchTableByColumnValue. Created ajaxes for rendering form and submitting of new List, and for rendering left side list upon adding a new list. Added text-overflow: ellipsis for the list to be shown partially on smaller devices. Also showing user name and avatar if it is in database